rwkv与rwkvcpp进行合并，使用strategy进行控制
