+++ README.md
@@ -143,6 +143,10 @@ Linux直接使用wenda环境执行 `python plugins/gen_data_st.py`
 
 默认参数在GTX1660Ti（6G显存）上正常运行，但速度较慢。
 
+设置strategy诸如"Q8_0->8"即支持量化在cpu运行，速度较慢，没有显卡或者没有nvidia显卡的用户使用。
+注意默认librwkv.so是在debian sid编译的，不支持windows,其他linux发行版本未知。可以查看：[saharNooby/rwkv.cpp](https://github.com/saharNooby/rwkv.cpp)，下载windows版本，或者自行编译。
+plugins/rwkvcpp里的三个模块来自[saharNooby/rwkv.cpp](https://github.com/saharNooby/rwkv.cpp)
+
 #### 生成小说
 ![](imgs/novel.png)
 #### 文字冒险游戏
@@ -154,10 +158,6 @@ Linux直接使用wenda环境执行 `python plugins/gen_data_st.py`
 模型位置等参数：修改`config.yml`。
 
 
-支持量化在cpu运行，速度较慢，没有显卡或者没有nvidia显卡的用户使用。
-注意默认librwkv.so是在debian sid编译的，不支持windows,其他linux发行版本未知。可以查看：[saharNooby/rwkv.cpp](https://github.com/saharNooby/rwkv.cpp)，下载windows版本，或者自行编译。
-
-plugins/rwkv_cpp_model.py,plugins/rwkv_cpp_shared_library.py,plugins/sampling.py这三个模块来自[saharNooby/rwkv.cpp](https://github.com/saharNooby/rwkv.cpp)
 
 ### llama
 运行：`run_llama.bat`。
