+++ example.config.yml
@@ -3,22 +3,15 @@ logging: False
 #日志"
 port: 17860
 #webui 默认启动端口号"
-llm_type: rwkvcpp
+llm_type: rwkv
 #llm模型类型:glm6b、rwkv、llama、replitcode
 llm_models: 
-  rwkvcpp: 
-     path: "/data/RWKV-4-Raven-7B-v11-Eng49%-Chn49%-Jpn1%-Other1%-20230430-ctx8192-Q8.bin"
-     strategy: "cuda fp16i8"
-     #rwkv模型位置"
-     historymode: string
-     #rwkv历史记录实现方式：state、string
-     cpu: 2
-     #设置使用cpu核的数量
   rwkv: 
      path: "model/rwkv-4-raven-7b-v11.pth"
      #rwkv模型位置"
      strategy: "cuda fp16i8"
      #rwkv模型参数"
+     #支持rwkvcpp运行方式，设置strategy诸如"Q8_0->16"即可开启，代表运行Q8_0模型在16个cpu核心上
      historymode: state
      #rwkv历史记录实现方式：state、string
   glm6b: 
@@ -31,6 +24,7 @@ llm_models:
      #cpu fp32	 所有glm模型 要直接跑在cpu上都可以使用这个参数
      #cpu fp16i8	fp16原生模型 要自行量化为int8跑在cpu上可以使用这个参数
      #cpu fp16i4	fp16原生模型要 自行量化为int4跑在cpu上可以使用这个参数"
+     #cuda:0 fp16 *14 -> cuda:1	fp16 多卡流水线并行，使用方法参考RWKV的strategy介绍。总层数28
      lora: ""
      #glm-lora模型位置
   llama:
