+++ plugins/rwkvcpp/librwkv.so
